Move Sibling class and wforce replication into separate source files
Previously Sibling and all replication code was all part of wforce.cc/hh
which was very confusing, with no clear modularization. This improves
the situation somewhat, and also allows e.g. Sibling to be used without
including all the wforce includes etc.
