+++ wforce/wforce.cc
@@ -47,6 +47,7 @@
 #include "wforce-web.hh"
 #include "twmap-wrapper.hh"
 #include "replication_sdb.hh"
+#include "wforce-replication.hh"
 #include "minicurl.hh"
 #include "iputils.hh"
 #include "ext/threadname.hh"
@@ -78,11 +79,6 @@ struct SiblingQueueItem {
   ComboAddress remote;
   std::shared_ptr<Sibling> recv_sibling;
 };
-
-static std::mutex g_sibling_queue_mutex;
-static std::queue<SiblingQueueItem> g_sibling_queue;
-static std::condition_variable g_sibling_queue_cv;
-size_t g_max_sibling_queue_size = 1000;
 
 std::string g_configDir; // where the config files are located
 std::shared_ptr<UserAgentParser> g_ua_parser_p;
@@ -425,115 +421,6 @@ void doConsole()
   }
 }
 
-Sibling::Sibling(const ComboAddress& ca) : Sibling(ca, Protocol::UDP)
-{
-}
-
-void Sibling::connectSibling()
-{
-  sockp = wforce::make_unique<Socket>(rem.sin4.sin_family, static_cast<int>(proto));
-  if (proto == Protocol::UDP) {
-    sockp->connect(rem);
-  }
-  else {
-    if (!d_ignoreself) {
-      try {
-        sockp->connect(rem);
-      }
-      catch (const NetworkError& e) {
-        errlog("TCP Connect to Sibling %s failed (%s)", rem
-                  .toStringWithPort(), e.what());
-      }
-    }
-  }
-}
-
-Sibling::Sibling(const ComboAddress& ca, Protocol p) : rem(ca), proto(p), d_ignoreself(false)
-{
-  if (!g_cmdLine.beClient) {
-    std::thread t([this]() {
-        thread_local bool init=false;
-        if (!init) {
-          setThreadName("wf/sibling-worker");
-          init = true;
-        }
-        connectSibling();
-        while (true) {
-          std::string msg;
-          {
-            std::unique_lock<std::mutex> lock(queue_mutx);
-            while (queue.size() == 0) {
-              queue_cv.wait(lock);
-            }
-            msg = std::move(queue.front());
-            queue.pop();
-          }
-          send(msg);
-        }
-      });
-    t.detach();
-  }
-}
-
-void Sibling::checkIgnoreSelf(const ComboAddress& ca)
-{
-  ComboAddress actualLocal;
-  actualLocal.sin4.sin_family = ca.sin4.sin_family;
-  socklen_t socklen = actualLocal.getSocklen();
-
-  if(getsockname(sockp->getHandle(), (struct sockaddr*) &actualLocal, &socklen) < 0) {
-    return;
-  }
-
-  actualLocal.sin4.sin_port = ca.sin4.sin_port;
-  if(actualLocal == rem) {
-    d_ignoreself=true;
-  }
-}
-
-void Sibling::send(const std::string& msg)
-{
-  if(d_ignoreself)
-    return;
-
-  if (proto == Protocol::UDP) {
-    if(::send(sockp->getHandle(),msg.c_str(), msg.length(),0) <= 0) {
-      ++failures;
-    }
-    else
-      ++success;
-  }
-  else if (proto == Protocol::TCP) {
-    // This needs protecting with a mutex because of the reconnect logic
-    std::lock_guard<std::mutex> lock(mutx);
-
-    uint16_t nsize = htons(msg.length());
-    try {
-      sockp->writen(std::string((char*)&nsize, sizeof(nsize)));
-      sockp->writen(msg);
-      ++success;
-    }
-    catch (const NetworkError& e) {
-      ++failures;
-      errlog("Error writing to Sibling %s, reconnecting (%s)", rem.toStringWithPort(), e.what());
-      connectSibling();
-    }
-  }
-}
-
-void Sibling::queueMsg(const std::string& msg)
-{
-  std::lock_guard<std::mutex> lock(queue_mutx);
-  if (queue.size() >= max_queue_size) {
-    errlog("Sibling::queueMsg: max sibling queue size (%d) reached - dropping replication msg");
-    return;
-  }
-  else {
-    queue.push(msg);
-  }
-  queue_cv.notify_one();
-}
-
 std::atomic<unsigned int> g_report_sink_rr(0);
 void sendReportSink(const LoginTuple& lt)
 {
@@ -566,39 +453,6 @@ void sendNamedReportSink(const std::string& msg)
 
     vec[j]->queueMsg(msg);
   }
-}
-
-
-GlobalStateHolder<vector<shared_ptr<Sibling>>> g_siblings;
-unsigned int g_num_sibling_threads = WFORCE_NUM_SIBLING_THREADS;
-SodiumNonce g_sodnonce;
-std::mutex sod_mutx;
-
-void encryptMsg(const std::string& msg, std::string& packet)
-{
-    std::lock_guard<std::mutex> lock(sod_mutx);
-    packet=g_sodnonce.toString();
-    packet+=sodEncryptSym(msg, g_key, g_sodnonce);
-}
-
-bool decryptMsg(const char* buf, size_t len, std::string& msg)
-{
-  SodiumNonce nonce;
-
-  if (len < static_cast<int>(crypto_secretbox_NONCEBYTES)) {
-    errlog("Could not decrypt replication operation: not enough bytes (%d) to hold nonce", len);
-    return false;
-  }
-  memcpy((char*)&nonce, buf, crypto_secretbox_NONCEBYTES);
-  string packet(buf + crypto_secretbox_NONCEBYTES, buf+len);
-  try {
-    msg=sodDecryptSym(packet, g_key, nonce);
-  }
-  catch (std::runtime_error& e) {
-    errlog("Could not decrypt replication operation: %s", e.what());
-    return false;
-  }
-  return true;
 }
 
 Json callWforceGetURL(const std::string& url, const std::string& password, std::string& err)
@@ -778,199 +632,6 @@ void checkSyncHosts()
   // If we didn't find a sync host we can't warm up
   if (found_sync_host == false)
     g_ping_up = true;
-}
-
-void replicateOperation(const ReplicationOperation& rep_op)
-{
-  auto siblings = g_siblings.getLocal();
-  string msg = rep_op.serialize();
-  string packet;
-
-  encryptMsg(msg, packet);
-
-  for(auto& s : *siblings) {
-    s->queueMsg(packet);
-  }
-}
-
-bool checkConnFromSibling(const ComboAddress& remote,
-                          shared_ptr<Sibling>& recv_sibling)
-{
-  auto siblings = g_siblings.getLocal();
-
-  for (auto &s : *siblings) {
-    if (ComboAddress::addressOnlyEqual()(s->rem, remote) == true) {
-      recv_sibling = s;
-      break;
-    }
-  }
-
-  if (recv_sibling == nullptr) {
-    errlog("Message received from host (%s) that is not configured as a sibling", remote.toStringWithPort());
-    return false;
-  }
-  return true;
-}
-
-void startReplicationWorkerThreads()
-{
-  for (size_t i=0; i<g_num_sibling_threads; i++) {
-    std::thread t([]() {
-        thread_local bool init=false;
-        if (!init) {
-          setThreadName("wf/repl-worker");
-          init = true;
-        }
-        while (true) {
-          SiblingQueueItem sqi;
-          {
-            std::unique_lock<std::mutex> lock(g_sibling_queue_mutex);
-            while (g_sibling_queue.size() == 0) {
-              g_sibling_queue_cv.wait(lock);
-            }
-            sqi = std::move(g_sibling_queue.front());
-            g_sibling_queue.pop();
-          }
-          ReplicationOperation rep_op;
-          if (rep_op.unserialize(sqi.msg) != false) {
-            rep_op.applyOperation();
-            sqi.recv_sibling->rcvd_success++;
-          }
-          else {
-            errlog("Invalid replication operation received from %s", sqi.remote.toString());
-            sqi.recv_sibling->rcvd_fail++;
-          }
-        }
-      });
-    t.detach();
-  }
-}
-
-void parseReceivedReplicationMsg(const std::string& msg, const ComboAddress& remote, std::shared_ptr<Sibling> recv_sibling)
-{
-  SiblingQueueItem sqi = { msg, remote, recv_sibling };
-  {
-    std::lock_guard<std::mutex> lock(g_sibling_queue_mutex);
-    if (g_sibling_queue.size() >= g_max_sibling_queue_size) {
-      errlog("parseReceivedReplicationMsg: max sibling queue size (%d) reached - dropping replication msg");
-      return;
-    }
-    else {
-      g_sibling_queue.push(sqi);
-    }
-  }
-  g_sibling_queue_cv.notify_one();
-}
-
-void parseTCPReplication(std::shared_ptr<Socket> sockp, const ComboAddress& remote, std::shared_ptr<Sibling> recv_sibling)
-{
-  setThreadName("wf/repl-tcp");
-
-  infolog("New TCP Replication connection from %s", remote.toString());
-  uint16_t size;
-  unsigned char ssize = sizeof(size);
-  char buffer[65535];
-  int len;
-  unsigned int num_rcvd=0;
-  
-  try {
-    while(true) {
-      len = sockp->read((char*)&size, ssize);
-      if (len != ssize) {
-        if (len)
-          errlog("parseTCPReplication: error reading size, got %d bytes, but was expecting %d bytes", len, ssize);
-        break;
-      }
-      size = ntohs(size);
-      if (!size) {
-        errlog("parseTCPReplication: Zero-length size field");
-        break;
-      }
-      if (size > sizeof(buffer)) {
-        errlog("parseTCPReplication: This should not happen - asked to read more than 65535 bytes");
-        break;
-      }
-      int size_read = 0;
-      if ((size_read = sockp->readAll(buffer, size)) != size) {
-        errlog("parseTCPReplication: Told to read %d bytes, but actually read %d bytes", size, size_read);
-        break;
-      }
-      std::string msg;
-      if (!decryptMsg(buffer, size_read, msg)) {
-        recv_sibling->rcvd_fail++;
-        continue;
-      }
-      parseReceivedReplicationMsg(msg, remote, recv_sibling);
-      num_rcvd++;
-    }
-    infolog("Received %d Replication entries from %s", num_rcvd, remote.toString());
-  }
-  catch(std::exception& e) {
-    errlog("ParseTCPReplication: client thread died with error: %s", e.what());
-  }
-}
-
-void receiveReplicationOperationsTCP(ComboAddress local)
-{
-  Socket sock(local.sin4.sin_family, SOCK_STREAM, 0);
-  ComboAddress remote=local;
-
-  setThreadName("wf/rcv-repl-tcp");
-  
-  sock.setReuseAddr();
-  sock.bind(local);
-  sock.listen(1024);
-  noticelog("Launched TCP sibling replication listener on %s", local.toStringWithPort());
-  
-  for (;;) {
-    // we wait for activity
-    try {
-      shared_ptr<Sibling> recv_sibling = nullptr;
-      std::shared_ptr<Socket> connp(sock.accept());
-      connp->setKeepAlive();
-      if (connp->getRemote(remote) && !checkConnFromSibling(remote, recv_sibling)) {
-        continue;
-      }
-      thread t1(parseTCPReplication, connp, remote, recv_sibling);
-      t1.detach();
-    }
-    catch(std::exception& e) {
-      errlog("receiveReplicationOperationsTCP: error accepting new connection: %s", e.what());
-    }
-  }
-}
-
-void receiveReplicationOperations(ComboAddress local)
-{
-  Socket sock(local.sin4.sin_family, SOCK_DGRAM);
-  sock.bind(local);
-  char buf[1500];
-  ComboAddress remote=local;
-  socklen_t remlen=remote.getSocklen();
-  int len;
-  auto siblings = g_siblings.getLocal();
-
-  setThreadName("wf/rcv-repl-udp");
-  
-  noticelog("Launched UDP sibling replication listener on %s", local.toStringWithPort());
-  for(;;) {
-    shared_ptr<Sibling> recv_sibling = nullptr;
-    len=recvfrom(sock.getHandle(), buf, sizeof(buf), 0, (struct sockaddr*)&remote, &remlen);
-    if(len <= 0 || len >= (int)sizeof(buf))
-      continue;
-
-    if (!checkConnFromSibling(remote, recv_sibling)) {
-      continue;
-    }
-    
-    string msg;
-    
-    if (!decryptMsg(buf, len, msg)) {
-      recv_sibling->rcvd_fail++;
-      continue;
-    }
-    parseReceivedReplicationMsg(msg, remote, recv_sibling);
-  }
 }
 
 AllowReturn defaultAllowTuple(const LoginTuple& lp)
